editor/libeditor/tests/test_bug1332876.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=795418
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 1332876</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1332876">Mozilla Bug 1332876</a>
<p id="display"></p>
<div id="content" style="display: none">
</div>

<iframe srcdoc="<html><body><span>Edit me!</span>"></iframe>

<pre id="test">

<script type="application/javascript">

/** Test for Bug 1332876 **/
SimpleTest.waitForExplicitFinish();
SimpleTest.waitForFocus(function() {
  var iframe = document.querySelector("iframe");
  iframe.contentDocument.designMode = "on";

  iframe.contentWindow.addEventListener("keypress", function() {
    iframe.style.display = "none";
    document.body.offsetHeight;
    ok(true, "did not crash");
    SimpleTest.finish();
  });

  iframe.contentWindow.addEventListener("click", function() {
    synthesizeKey("a", {}, iframe.contentWindow);
  });

  synthesizeMouse(iframe, 20, 20, {});
});

</script>

</pre>
</body>
</html>
